requirements.txt
cffi==1.16.0
contourpy==1.2.0
cssselect2==0.7.0
cycler==0.12.1
defusedxml==0.7.1
fonttools==4.49.0
numpy==1.26.4
packaging==24.0
pillow==10.2.0
pycparser==2.21
pyparsing==3.1.2
python-dateutil==2.9.0.post0
six==1.16.0
tinycss2==1.2.1
webencodings==0.5.1
